Adicionado Horários

+++ src/main/java/com/example/qts/discipine/DisciplineRepository.java
@@ -2,5 +2,5 @@ package com.example.qts.discipine;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DisciplineRepository extends JpaRepository <Discipline, Integer> {
+public interface DisciplineRepository extends JpaRepository <Discipline, Long> {
 }
